Move refactored difficulty calculator from newCalc to oppai
